Add "Archived Recordings" view tab (#69)
* Extract active recording list component

* Add Archived Recordings view

Add Archived Recordings view under Recordings List tab, visible/enabled
only when the ContainerJFR instance supports archiving

* Add "Archive" button to active recordings view

Add button allowing users to select active recordings and save them to
the archives

Fixes #57

* Immediately refresh archived list when save message observed

* Refactor to extract shared data list table element

* Use PF DataToolbar instead of plain Toolbar
