+++ src/app/RecordingList/RecordingList.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import { useHistory, useRouteMatch } from 'react-router-dom';
-import { filter, map } from 'rxjs/operators';
-import { Button, Card, CardBody, CardHeader, DataList, DataListCheck, DataListItem, DataListItemRow, DataListItemCells, DataListCell, PageSection, Text, TextVariants, Title, Toolbar, ToolbarGroup, ToolbarItem } from '@patternfly/react-core';
+import { Card, CardBody, CardHeader, Tabs, Tab, Text, TextVariants } from '@patternfly/react-core';
 import { ServiceContext } from '@app/Shared/Services/Services';
 import { TargetView } from '@app/TargetView/TargetView';
+import { ActiveRecordingsList } from './ActiveRecordingsList';
+import { ArchivedRecordingsList } from './ArchivedRecordingsList';
 
-interface Recording {
+export interface Recording {
   id: number;
   name: string;
   state: RecordingState;
@@ -19,7 +19,7 @@ interface Recording {
   reportUrl: string;
 }
 
-enum RecordingState {
+export enum RecordingState {
   STOPPED = 'STOPPED',
   STARTING = 'STARTING',
   RUNNING = 'RUNNING',
@@ -28,179 +28,35 @@ enum RecordingState {
 
 export const RecordingList = (props) => {
   const context = React.useContext(ServiceContext);
-  const routerHistory = useHistory();
-
-  const [recordings, setRecordings] = React.useState([]);
-  const [headerChecked, setHeaderChecked] = React.useState(false);
-  const [checkedIndices, setCheckedIndices] = React.useState([] as number[]);
-  const { path, url } = useRouteMatch();
-
-  const tableColumns: string[] = [
-    'Name',
-    'Start Time',
-    'Duration',
-    'Download',
-    'Report',
-    'State',
-  ];
-
-  const handleCreateRecording = () => {
-    routerHistory.push(`${url}/create`);
-  };
-
-  const handleHeaderCheck = (checked) => {
-    setHeaderChecked(checked);
-    setCheckedIndices(checked ? recordings.map((r, idx) => idx) : []);
-  };
-
-  const handleRowCheck = (checked, index) => {
-    if (checked) {
-      setCheckedIndices(ci => ([...ci, index]));
-    } else {
-      setHeaderChecked(false);
-      setCheckedIndices(ci => ci.filter(v => v !== index));
-    }
-  };
-
-  const handleDeleteRecordings = () => {
-    recordings.forEach((r: Recording, idx) => {
-      if (checkedIndices.includes(idx)) {
-        handleRowCheck(false, idx);
-        context.commandChannel.sendMessage('delete', [ r.name ]);
-      }
-    });
-    context.commandChannel.sendMessage('list');
-  };
-
-  const handleStopRecordings = () => {
-    recordings.forEach((r: Recording, idx) => {
-      if (checkedIndices.includes(idx)) {
-        handleRowCheck(false, idx);
-        if (r.state === RecordingState.RUNNING || r.state === RecordingState.STARTING) {
-          context.commandChannel.sendMessage('stop', [ r.name ]);
-        }
-      }
-    });
-    context.commandChannel.sendMessage('list');
-  };
+  const [activeTab, setActiveTab] = React.useState(0);
+  const [archiveEnabled, setArchiveEnabled] = React.useState(false);
 
   React.useEffect(() => {
-    const sub = context.commandChannel.onResponse('list')
-      .pipe(
-        filter(m => m.status === 0),
-        map(m => m.payload),
-      )
-      .subscribe(recordings => setRecordings(recordings));
+    const sub = context.commandChannel.isArchiveEnabled().subscribe(enabled => setArchiveEnabled(enabled));
     return () => sub.unsubscribe();
   }, []);
-
-  React.useEffect(() => {
-    context.commandChannel.sendMessage('list');
-    const id = setInterval(() => context.commandChannel.sendMessage('list'), 5000);
-    return () => clearInterval(id);
-  }, []);
-
-  const RecordingRow = (props) => {
-    return (
-      <DataListItemRow>
-        <DataListCheck aria-labelledby="table-row-1-1" name={`row-${props.index}-check`} onChange={(checked) => handleRowCheck(checked, props.index)} isChecked={checkedIndices.includes(props.index)} />
-        <DataListItemCells
-          dataListCells={[
-            <DataListCell key={`table-row-${props.index}-1`}>
-              {props.recording.name}
-            </DataListCell>,
-            <DataListCell key={`table-row-${props.index}-2`}>
-              <ISOTime timeStr={props.recording.startTime} />
-            </DataListCell>,
-            <DataListCell key={`table-row-${props.index}-3`}>
-              <RecordingDuration duration={props.recording.duration} />
-            </DataListCell>,
-            <DataListCell key={`table-row-${props.index}-4`}>
-              <Link url={`${props.recording.downloadUrl}.jfr`} />
-            </DataListCell>,
-            // TODO make row expandable and render report in collapsed iframe
-            <DataListCell key={`table-row-${props.index}-5`}>
-              <Link url={props.recording.reportUrl} />
-            </DataListCell>,
-            <DataListCell key={`table-row-${props.index}-6`}>
-              {props.recording.state}
-            </DataListCell>
-          ]}
-        />
-      </DataListItemRow>
-    );
-  };
-
-  const ISOTime = (props) => {
-    const fmt = new Date(props.timeStr).toISOString();
-    return (<span>{fmt}</span>);
-  };
-
-  const RecordingDuration = (props) => {
-    const str = props.duration === 0 ? 'Continuous' : `${props.duration / 1000}s`
-    return (<span>{str}</span>);
-  };
-
-  const Link = (props) => {
-    return (<a href={props.url} target="_blank">{props.display || props.url}</a>);
-  };
-
-  const isStopDisabled = () => {
-    if (!checkedIndices.length) {
-      return true;
-    }
-    const filtered = recordings.filter((r: Recording, idx: number) => checkedIndices.includes(idx));
-    const anyRunning = filtered.some((r: Recording) => r.state === RecordingState.RUNNING || r.state == RecordingState.STARTING);
-    return !anyRunning;
-  };
-
-  const RecordingsToolbar = (props) => {
-    return (
-      <Toolbar>
-        <ToolbarGroup>
-          <ToolbarItem>
-            <Button variant="primary" onClick={handleCreateRecording}>Create</Button>
-          </ToolbarItem>
-        </ToolbarGroup>
-        <ToolbarGroup>
-          <ToolbarItem>
-            <Button variant="secondary" onClick={handleStopRecordings} isDisabled={isStopDisabled()}>Stop</Button>
-          </ToolbarItem>
-        </ToolbarGroup>
-        <ToolbarGroup>
-          <ToolbarItem>
-            <Button variant="danger" onClick={handleDeleteRecordings} isDisabled={!checkedIndices.length}>Delete</Button>
-          </ToolbarItem>
-        </ToolbarGroup>
-      </Toolbar>
-    );
-  };
 
   return (
     <TargetView pageTitle="Recordings">
       <Card>
-        <CardHeader><Text component={TextVariants.h4}>Active Recordings</Text></CardHeader>
         <CardBody>
-          <RecordingsToolbar />
-          <DataList aria-label="Recording List">
-            <DataListItem aria-labelledby="table-header-1">
-              <DataListItemRow>
-                <DataListCheck aria-labelledby="table-header-1" name="header-check" onChange={handleHeaderCheck} isChecked={headerChecked} />
-                <DataListItemCells
-                  dataListCells={tableColumns.map((key , idx) => (
-                    <DataListCell key={key}>
-                      <span id={`table-header-${idx}`}>{key}</span>
-                    </DataListCell>
-                  ))}
-                />
-              </DataListItemRow>
-            </DataListItem>
-            <DataListItem aria-labelledby="table-row-1-1">
-            {
-              recordings.map((r, idx) => <RecordingRow recording={r} index={idx}/>)
-            }
-            </DataListItem>
-          </DataList>
+          {
+            archiveEnabled ? (
+              <Tabs activeKey={activeTab} onSelect={(evt, idx) => setActiveTab(Number(idx))}>
+                <Tab eventKey={0} title="Active Recordings">
+                  <ActiveRecordingsList archiveEnabled={true}/>
+                </Tab>
+                <Tab eventKey={1} title="Archived Recordings">
+                  <ArchivedRecordingsList />
+                </Tab>
+              </Tabs>
+            ) : (
+              <>
+                <CardHeader><Text component={TextVariants.h4}>Active Recordings</Text></CardHeader>
+                <ActiveRecordingsList archiveEnabled={false}/>
+              </>
+            )
+          }
         </CardBody>
       </Card>
     </TargetView>
